gamma/requirements.txt
asgiref==3.5.0
cycler==0.11.0
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
fonttools==4.29.1
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.2
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
